ignored qemu fuzzer for non-linux
